requirements.txt
dacite==1.5.0
flask==1.1.1
Flask-Injector==0.12.3
flask-restplus==0.13.0
flask_jwt_extended==3.24.1
google-cloud-firestore==1.6.2
injector==0.18.3
marshmallow==3.5.0
werkzeug==0.16.1
